+++ packages/transaction-pay-controller/src/strategy/relay/types.ts
@@ -56,6 +56,17 @@ export type RelayQuote = {
     relayer: {
       amountUsd: string;
     };
+    subsidized?: {
+      amount: string;
+      amountFormatted: string;
+      amountUsd: string;
+      currency: {
+        address: Hex;
+        chainId: number;
+        decimals: number;
+      };
+      minimumAmount: string;
+    };
   };
   metamask: {
     gasLimits: number[];